requirements.txt
python-dotenv==0.10.5
python-telegram-bot==12.3.0
requests==2.22.0
dialogflow==0.7.2